Adjust whitespace, change styling rules.
I also changed the project container divs to unordered lists's and the project paragraphs to list items because it makes more sense semantically.

The CSS rules are easier to read, better for accessibility, and add a fun markdown-esque thing to the project lists. I also added a bit more internal docs to everything.